Move app_state, action, serializers to models dir
