+++ lib/models/problem.dart
@@ -5,8 +5,8 @@ import 'dart:convert';
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import 'package:podcustard/redux/app_state.dart';
-import 'package:podcustard/redux/serializers.dart';
+import 'package:podcustard/models/app_state.dart';
+import 'package:podcustard/models/serializers.dart';
 
 part 'problem.g.dart';
