+++ lib/models/provider_info.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import 'package:podcustard/redux/serializers.dart';
+import 'package:podcustard/models/serializers.dart';
 
 part 'provider_info.g.dart';
