+++ lib/models/user.dart
@@ -6,7 +6,7 @@ import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
 import 'package:podcustard/models/provider_info.dart';
-import 'package:podcustard/redux/serializers.dart';
+import 'package:podcustard/models/serializers.dart';
 
 part 'user.g.dart';
