+++ lib/widgets/app.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart' hide Action;
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
 import 'package:podcustard/models/user.dart';
-import 'package:podcustard/redux/actions.dart';
-import 'package:podcustard/redux/app_state.dart';
+import 'package:podcustard/models/actions.dart';
+import 'package:podcustard/models/app_state.dart';
 import 'package:podcustard/widgets/auth_page.dart';
 import 'package:podcustard/widgets/main_page.dart';
