+++ test/unit/middleware_test.dart
@@ -2,9 +2,9 @@ import 'package:mockito/mockito.dart';
 import 'package:podcustard/models/problem.dart';
 import 'package:redux/redux.dart';
 import 'package:podcustard/models/user.dart';
-import 'package:podcustard/redux/actions.dart';
+import 'package:podcustard/models/actions.dart';
 import 'package:podcustard/redux/app_reducer.dart';
-import 'package:podcustard/redux/app_state.dart';
+import 'package:podcustard/models/app_state.dart';
 import 'package:podcustard/redux/middleware.dart';
 import 'package:podcustard/services/auth_service.dart';
 import 'package:test/test.dart';
